change base route to /, fix dmoj api access

+++ dmojsolutions/views/admin.py
@@ -5,8 +5,8 @@ from urllib.error import URLError
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.core.exceptions import ObjectDoesNotExist
+from django.db import transaction
 from django.shortcuts import redirect
-from django.urls import reverse
 
 from dmojsolutions.models import Problem, Solution
 from view_utils import superuser_required
@@ -18,7 +18,7 @@ def get_json(url):
         return json.loads(resp.read())
 
 
-PROBLEMS_LIST_URL = 'https://dmoj.ca/api/problem/list'
+PROBLEMS_LIST_URL = 'https://dmoj.ca/api/v2/problems'
 
 
 @superuser_required(url_pattern_name='dmojsolutions:index')
@@ -26,12 +26,28 @@ PROBLEMS_LIST_URL = 'https://dmoj.ca/api/problem/list'
 def reload_problems(req):
     try:
         new_problems = get_json(PROBLEMS_LIST_URL)
+        total_pages = new_problems['data']['total_pages']
 
-        Problem.objects.all().delete()
+        messages.add_message(req, messages.INFO, 'Found %d pages of problems!' % total_pages)
+
         problems = []
-        for k, obj in new_problems.items():
-            problems.append(Problem(code=k, name=obj['name'], group=obj['group']))
-        Problem.objects.bulk_create(problems)
+        for page in range(1, total_pages + 1):
+            if page == 1:
+                page_problems = new_problems
+            else:
+                page_problems = get_json(f'{PROBLEMS_LIST_URL}?page={page}')
+
+            for obj in page_problems['data']['objects']:
+                problems.append(Problem(code=obj['code'], name=obj['name'], group=obj['group']))
+
+            messages.add_message(req, messages.INFO, 'Loaded %d problems from page %d!' % (len(page_problems['data']['objects']), page))
+
+        if not problems:
+            raise ValueError('Problem API returned no objects.')
+
+        with transaction.atomic():
+            Problem.objects.all().delete()
+            Problem.objects.bulk_create(problems)
 
         messages.add_message(req, messages.INFO, 'Added %d problems!' % len(problems))
     except URLError as e:
@@ -39,8 +55,9 @@ def reload_problems(req):
             msg = 'Error %s: %s' % (e.code, e.reason)
         else:
             msg = '%s' % e.reason
-
         messages.add_message(req, messages.WARNING, msg)
+    except (KeyError, TypeError, ValueError) as e:
+        messages.add_message(req, messages.WARNING, f'Problem import failed: {e}')
 
     return redirect('dmojsolutions:index')
